Extract subselection tests to separate spec
